refactor: rename fern-sidebar to SidebarBuilder
