+++ doc-gen/src/SidebarBuilder.ts
